Maintain results
When reconciling Versions, check for the label "runId". If found, create
a Results with that name that includes a test item for each test ID
created.

Add a TestPods controller that watches for changes to the phase of the
test pods and updates the Results objects.

Add a Results controller that runs on create to update each test pod
result. This is to handle the race condition where the test pod finishes
before the Results is created, which means the TestPods controller would
not have updated the Results.